COMP: Took care of the last few compiler warnings in the 2D chart code.

+++ VTK/Charts/vtkBlockItem.cxx
@@ -103,8 +103,8 @@ bool vtkBlockItem::MouseEnterEvent(const vtkContextMouseEvent &)
 //-----------------------------------------------------------------------------
 bool vtkBlockItem::MouseMoveEvent(const vtkContextMouseEvent &mouse)
 {
-  int deltaX = mouse.Pos[0] - this->LastPosition[0];
-  int deltaY = mouse.Pos[1] - this->LastPosition[1];
+  float deltaX = mouse.Pos[0] - this->LastPosition[0];
+  float deltaY = mouse.Pos[1] - this->LastPosition[1];
   this->LastPosition[0] = mouse.Pos[0];
   this->LastPosition[1] = mouse.Pos[1];
